Script Enhancements Based on Testing Feedback

+++ toPolymer3.sh
@@ -1,4 +1,3 @@
-#!/bin/bash
 
 if [ "$1" != "" ]; then
 	echo "*** Commit Message: $1 ***"
@@ -9,6 +8,10 @@ fi
 
 echo "*** Make Polymer 3 branch ***"
 git checkout -b polymer-3-`date +%Y-%m-%d-%H%M%S` master
+
+echo "*** Remove index.html and all-imports.html ***"
+rm -f index.html
+rm -f all-imports.html
 
 echo "*** Temporarily changing tabs to spaces for all js and html files ***"
 array=(`find . \( -name "*.html" -o -name "*.js" \) -not \( -path "./node_modules*" -o -path "./bower_components*" -o -path "./.git*" \)`)
@@ -23,7 +26,7 @@ echo "toPolymer3.sh" >> ./.gitignore
 
 echo "*** Committing this change as the modulizer needs a clean repo ***"
 git add -u
-git commit -m "Temporarily change tabs to spaces for modulizer"
+git commit -m "Cleanup files and temporarily change tabs to spaces for modulizer"
 
 echo "*** Run bower install ***"
 rm -rf bower_components
@@ -33,11 +36,14 @@ npm install -g bower
 bower install
 
 echo "*** Remove node_modules folder (Windows issue, otherwise you get a heap stack error) ***"
-rm -r node_modules
+rm -rf node_modules
+
+echo "*** Remove package-lock.json file ***"
+rm -f package-lock.json
 
 echo "*** Run the modulizer ***"
 npm install -g polymer-modulizer
-modulizer --out .
+echo | modulizer --out .
 
 echo "*** Discard the travis file changes (they are not needed) ***"
 git checkout .travis.yml
@@ -118,34 +124,52 @@ sed -i.original '/\"lint\":/c\
 sed -i.original '/\"lint:html\":/c\
 \    \"lint:js\": \"eslint . test/** demo/** --ext .js,.html\",
 ' package.json
+sed -i.original '/\"lint:wc\":/c\
+\    \"lint:wc\": \"polymer lint\",
+' package.json
 sed -i.original '/\"test:lint\":/c\
 \    \"test:lint\": \"npm run test:lint:wc && npm run test:lint:js\",
 ' package.json
 sed -i.original '/\"test:lint:html\":/c\
 \    \"test:lint:js\": \"eslint . test/** demo/** --ext .js,.html\",
 ' package.json
+sed -i.original '/\"test:lint:wc\":/c\
+\    \"test:lint:wc\": \"polymer lint\",
+' package.json
 rm -f package.json.original
 
 echo "*** Convert d2l bower components to polymer-3 npm versions ***"
 declare -A dependencies
-dependencies["d2l-colors"]="git+https://github.com/BrightspaceUI/colors.git#polymer-3"
-dependencies["d2l-typography"]="git+https://github.com/BrightspaceUI/typography.git#polymer-3"
-dependencies["d2l-offscreen"]="git+https://github.com/BrightspaceUI/offscreen.git#polymer-3"
-dependencies["d2l-polymer-behaviors"]="git+https://github.com/Brightspace/d2l-polymer-behaviors-ui.git#polymer-3.x"
-dependencies["d2l-icons"]="git+https://github.com/BrightspaceUI/icons.git#polymer-3"
-dependencies["d2l-button"]="git+https://github.com/BrightspaceUI/button.git#polymer-3"
-dependencies["d2l-link"]="git+https://github.com/BrightspaceUI/link.git#polymer-3"
+dependencies=(
+["d2l-colors"]="git+https://github.com/BrightspaceUI/colors.git#polymer-3"
+["d2l-typography"]="git+https://github.com/BrightspaceUI/typography.git#polymer-3"
+["d2l-offscreen"]="git+https://github.com/BrightspaceUI/offscreen.git#polymer-3"
+["d2l-polymer-behaviors"]="git+https://github.com/Brightspace/d2l-polymer-behaviors-ui.git#polymer-3.x"
+["d2l-icons"]="git+https://github.com/BrightspaceUI/icons.git#polymer-3"
+["d2l-button"]="git+https://github.com/BrightspaceUI/button.git#polymer-3"
+["d2l-link"]="git+https://github.com/BrightspaceUI/link.git#polymer-3"
+)
 
 array=(`grep -i "^    \"d2l-.*\":" bower.json | cut -d"\"" -f2`)
+depError=false
+errorComponents=""
 for line in "${array[@]}"
 do
 	echo $line
 	if [ ${dependencies[$line]} ]
 	then
 		echo ${dependencies[$line]}
-		npm i --save --package-lock-only --no-package-lock ${dependencies[$line]}
+		npm i --save --package-lock-only --no-package-lock ${dependencies[$line]} || { errorComponents="$errorComponents * $line"; depError=true; } 
 	fi
 done
+if $depError
+then
+	echo ""
+	echo "*** FAILURE ***"
+	echo "The following components have not been migrated to Polymer 3 yet and should be done first:"
+	echo "$errorComponents"
+	exit 1;
+fi
 
 echo "*** Remove bower_components directory and bower.json ***"
 rm -rf bower_components
@@ -179,3 +203,6 @@ git reset -- toPolymer3.sh
 git reset -- README.md
 git checkout README.md
 git commit -m "Polymer 3 Conversion $message"
+
+echo "*** Re-install npm dependencies ***"
+npm i
